refactor: minor changes after cypress 10 update
