+++ test/cypress/README.md
@@ -55,7 +55,7 @@ Other remote environments could also be used (see `e2e/config/*.json`).
 2. Create (or update) `.env` file with the correct paths (see `template.env`)
 
 3. To start localtest, app frontend, and the app you configured above, run the command below.
-   (Skip to step 4 if the solutions are already running). The command will not finish, but it will start the app-frontend server in development mode. When you see the message `[app-frontend] No errors found.` you can move on to the next step.
+   (Skip to step 4 if the solutions are already running). The command will not finish, but it will start the app-frontend server in development mode. This command may take some time, depending on if LocalTest has been setup earlier or not, and if the docker cache is hit or not. When the output of this command seems to have stopped, you can continue to the next step.
 
 ```cmd
     yarn run before:appfrontend
