+++ test/cypress/cypress.config.js
@@ -1,27 +1,24 @@
 const { defineConfig } = require('cypress');
-const path = require('path');
-const fs = require('fs-extra');
+const path = require('node:path');
+const fs = require('node:fs/promises');
 
 module.exports = defineConfig({
   env: {
     multiData2Stage: 'frontend-test',
     stateless: 'stateless-app',
-    anonymous: 'anonymous-stateless-app'
+    anonymous: 'anonymous-stateless-app',
   },
   e2e: {
     setupNodeEvents(on, config) {
-      var pathToConfig = 'e2e/config';
-      switch (config.env.environment) {
-        case 'local':
-          return getConfigurationByFile(pathToConfig, 'local');
-        case 'at21':
-          return getConfigurationByFile(pathToConfig, 'at21');
-        case 'at22':
-          return getConfigurationByFile(pathToConfig, 'at22');
-        case 'tt02':
-          return getConfigurationByFile(pathToConfig, 'tt02');
+      const validEnvironments = ['local', 'at21', 'at22', 'tt02'];
+
+      if (validEnvironments.includes(config.env.environment)) {
+        return getConfigurationByFile(config.env.environment);
       }
-      return config;
+
+      throw new Error(`Unknown environment "${config.env.environment}"
+Valid environments are:
+- ${validEnvironments.join('\n- ')}`);
     },
     specPattern: 'e2e/integration/',
     supportFile: 'e2e/support/index.js',
@@ -39,16 +36,15 @@ module.exports = defineConfig({
   defaultCommandTimeout: 8000,
   reporter: 'junit',
   reporterOptions: {
-    mochaFile: 'reports/result-[hash].xml'
+    mochaFile: 'reports/result-[hash].xml',
   },
   retries: {
     runMode: 1,
-    openMode: 0
-  }
+    openMode: 0,
+  },
 });
 
-function getConfigurationByFile(pathToFile, file) {
-  const pathToJsonDataFile = path.resolve(pathToFile, `${file}.json`);
-  return fs.readJson(pathToJsonDataFile);
+async function getConfigurationByFile(file) {
+  const pathToJsonDataFile = path.resolve('e2e/config', `${file}.json`);
+  return JSON.parse((await fs.readFile(pathToJsonDataFile)).toString());
 }
-
